% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.R
\name{bciex}
\alias{bciex}
\title{bciex: Data sets for examples from BCI.}
\section{Installation}{
\preformatted{# install.packages("remotes")
remotes::install_github("forestgeo/bciex")
}

\href{https://forestgeo.github.io/bciex/}{Learn more about bciex}
}

