% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.R
\name{BIOMASS}
\alias{BIOMASS}
\title{BIOMASS: Estimate above-ground biomass in tropical forests.}
\description{
\itemize{
\item \href{https://CRAN.R-project.org/package=BIOMASS}{Description}
\item \href{https://cran.r-project.org/web/packages/BIOMASS/BIOMASS.pdf}{Manual}
\item \href{https://cran.r-project.org/web/packages/BIOMASS/vignettes/VignetteBiomass.html}{Vignette}
}
}
\section{Installation}{
\preformatted{install.packages("BIOMASS")
}

\href{https://CRAN.R-project.org/package=BIOMASS}{Learn more about BIOMASS}
}

