% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo-package.R
\docType{package}
\name{fgeo-package}
\alias{fgeo}
\alias{fgeo-package}
\title{fgeo: Analyze ForestGEO data.}
\description{
fgeo: Analyze ForestGEO data.
}
\examples{
fgeo_link <- function(x) {
  x \%>\%
    dplyr::mutate(
      topic = glue::glue_data(
        .,
        "<a href=https://forestgeo.github.io/\\\\
        {package}/reference/{topic}>{topic}</a>"
      ),
      package = glue::glue_data(
        .,
        "<a href=https://forestgeo.github.io/\\\\
        {package}>{package}</a>"
      )
      ) \%>\%
    dplyr::arrange(package)
}

fgeo_docs("", package, topic, alias, title, keyword) \%>\%
  dplyr::filter(!keyword \%in\% "internal") \%>\%
  dplyr::select(.data$alias, .data$topic, dplyr::everything()) \%>\%
  fgeo_link() \%>\%
  dplyr::select(-.data$package, -.data$keyword) \%>\%
  dplyr::rename(name = .data$alias, source = .data$topic)
}
\seealso{
Useful links:
\itemize{
  \item \url{http://forestgeo.github.io/fgeo}
  \item \url{https://github.com/forestgeo/fgeo}
  \item Report bugs at \url{https://github.com/forestgeo/fgeo/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} [contractor]

Authors:
\itemize{
  \item Gabriel Arellano \email{gabriel.arellano.torres@gmail.com} [reviewer]
  \item  Richard  Condit \email{richardcondit@gmail.com}
  \item Stuart Davies \email{daviess@si.edu} [reviewer]
  \item Matteo Detto \email{teoparaglider@gmail.com}
  \item Pamela Hall \email{phall@alum.mit.edu}
  \item Kyle Harms \email{ kharms@lsu.edu}
  \item David Kenfack \email{KenfackD@si.edu } [reviewer]
  \item Suzanne Lao \email{laoz@si.edu} [reviewer]
  \item Sean McMahon \email{mcmahons@si.edu} [reviewer]
  \item Sabrina Russo \email{srusso2@unl.edu} [reviewer]
  \item Kristina Teixeira \email{TeixeiraK@si.edu} [reviewer]
  \item Graham Zemunik \email{grah.zem@gmail.com} [reviewer]
  \item Daniel Zuleta \email{dfzuleta@gmail.com} [reviewer]
}

Other contributors:
\itemize{
  \item Lauren Krizel \email{KrizelL@si.edu} [reviewer]
  \item Haley Overstreet \email{OverstreetH@si.edu} [reviewer]
  \item CTFS-ForestGEO \email{ForestGEO@si.edu} [copyright holder, funder]
}

}
