% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.R
\name{fgeo.demography}
\alias{fgeo.demography}
\title{fgeo.demography: Analyze recruitment, mortality and growth.}
\section{Installation}{
\preformatted{# install.packages("remotes")
remotes::install_github("forestgeo/fgeo.demography")
}

\href{https://forestgeo.github.io/fgeo.demography/}{Learn more about fgeo.demography}
}

