% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.R
\name{fgeo.map}
\alias{fgeo.map}
\title{fgeo.map: Map ForestGEO's data.}
\section{Installation}{
\preformatted{# install.packages("remotes")
remotes::install_github("forestgeo/fgeo.map")
}

\href{https://forestgeo.github.io/fgeo.map/}{Learn more about fgeo.map}
}

