% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.R
\name{fgeo.tool}
\alias{fgeo.tool}
\title{fgeo.tool: ForestGEO's tools for general purposes.}
\section{Installation}{

To install \strong{fgeo.tool} run the following code in R:\preformatted{missing_remotes <- !require(remotes)
if (missing_remotes) {
  install.packages("remotes")
}
remotes::install_github("forestgeo/fgeo.tool")
}
\itemize{
\item To learn the details about how to install packages from GitHub, read
\href{https://goo.gl/dQKEeg}{this blog post}.
\item \href{https://forestgeo.github.io/fgeo.tool/}{Go to the website of fgeo.tool}
}
}

