% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_browse.R
\name{fgeo_browse}
\alias{fgeo_browse}
\alias{fgeo_browse_reference}
\title{Open a web browser on fgeo's website.}
\usage{
fgeo_browse()

fgeo_browse_reference()
}
\description{
Load fgeo's URLs into an HTML browser.
}
\examples{
if (interactive()) {
  fgeo_browse()
  fgeo_browse_reference()
}
}
\seealso{
Other functions to explore fgeo: \code{\link{fgeo_help}},
  \code{\link{fgeo_packages}}
}
\concept{functions to explore fgeo}
