% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{fgeo_conflicts}
\alias{fgeo_conflicts}
\title{Conflicts between the fgeo and other packages.}
\usage{
fgeo_conflicts()
}
\description{
This function lists all the conflicts between packages in fgeo
and other packages that you have loaded.
}
\examples{
fgeo_conflicts()
}
