% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{fgeo_core}
\alias{fgeo_core}
\title{Core packages of \strong{fgeo}.}
\usage{
fgeo_core()
}
\value{
A string.
}
\description{
Core packages of \strong{fgeo}.
}
\seealso{
Other functions for developers: \code{\link{fgeo_dependencies}},
  \code{\link{fgeo_docs}}
}
\concept{functions for developers}
\keyword{internal}
