% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_dependencies.R
\name{fgeo_dependencies}
\alias{fgeo_dependencies}
\title{List package dependencies (imports and suggests) of \strong{fgeo}.}
\usage{
fgeo_dependencies(include_self = TRUE, section = c("Imports",
  "Suggests"))
}
\arguments{
\item{include_self}{Include fgeo in the list?}

\item{section}{Either "Imports" or "Suggests".}
}
\description{
List package dependencies (imports and suggests) of \strong{fgeo}.
}
\examples{
fgeo_dependencies(TRUE, "Imports")
}
