% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{fgeo_deps}
\alias{fgeo_deps}
\title{List all fgeo dependencies}
\usage{
fgeo_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
fgeo packages.}
}
\description{
List all fgeo dependencies
}
