% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_docs.R
\name{fgeo_docs}
\alias{fgeo_docs}
\title{Explore the documentation of \strong{fgeo}.}
\usage{
fgeo_docs(pattern = NULL, ..., package = fgeo::fgeo_core())
}
\arguments{
\item{pattern}{Character string of length one to match an entry in packages
documentation.}

\item{...}{Bare names of the columns to select. Valid names are \code{package},
\code{name}, \code{title}, \code{topic}, \code{type}, \code{alias}, \code{keyword}, \code{concept}. If no name
is given, then all names are returned.}

\item{package}{A character vector with the names of packages to search
through, or NULL in which case all available packages in the library trees
specified by lib.loc are searched. Defaults to search only the core
packages of \strong{fgeo}.}
}
\value{
A dataframe.
}
\description{
Explore the documentation of \strong{fgeo}.
}
\examples{
# Filter rows with a matching pattern.
fgeo_docs("predicate")

# Select specific columns
fgeo_docs("predicate", concept, topic, title)

# Exclude specific columns
fgeo_docs("abundance", -package)
}
\seealso{
Other functions for developers: \code{\link{fgeo_dependencies}},
  \code{\link{fgeo}}
}
\concept{functions for developers}
\keyword{internal}
