% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_docs_addin.R
\name{fgeo_docs_addin}
\alias{fgeo_docs_addin}
\title{Search the documentation of \strong{fgeo}.}
\usage{
fgeo_docs_addin()
}
\description{
\code{fgeo_docs_addin()} opens an \href{https://shiny.rstudio.com/articles/gadgets.html}{RStudio gadget} and
\href{http://rstudio.github.io/rstudioaddins/}{addin} that allows you to search
\strong{fgeo}'s documentation. Appears as "fgeo_docs" in the RStudio Addins menu.
}
