% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_functions.R
\name{fgeo_functions}
\alias{fgeo_functions}
\title{Overview fgeo's functions.}
\usage{
fgeo_functions()
}
\value{
A dataframe.
}
\description{
Overview fgeo's functions.
}
\section{Acknowledgements}{

Thanks to Jim Hester for advice on an implementation detail
(https://goo.gl/Qzp5kR).
}

\examples{
head(fgeo_functions())
}
