% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_help.R
\name{fgeo_help}
\alias{fgeo_help}
\title{Search \strong{fgeo} for documentation matching a given character string.}
\usage{
fgeo_help(pattern = "", package = fgeo::fgeo_core())
}
\arguments{
\item{pattern}{A character string to be matched in the name, alias or title
of a topic's documentation. Defaults to matching anything, which retrieves
all the documentation of \strong{fgeo} packages.}

\item{package}{A character string giving the name or one or more
core-packages of \strong{fgeo}.}
}
\value{
The results are returned in a list object of class "hsearch", which
has a print method for nicely formatting the results of the query.
}
\description{
Use this function with a string to match the \emph{name} of a topic (generally the
name of a function or dataset), or text from the \emph{title} or \emph{alias} sections
of the topic's documentation.
}
\examples{
if (interactive()) fgeo_help()
if (interactive()) fgeo_help("fgeo", package = "fgeo")

results <- fgeo_help("fgeo", package = "fgeo.tool")
if (interactive()) results
str(results)
}
\seealso{
\code{\link[utils:help.search]{utils::help.search()}}, \code{\link[fgeo:fgeo_core]{fgeo::fgeo_core()}}.
}
