% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_packages.R
\name{fgeo_index}
\alias{fgeo_index}
\title{Create a table of some component of \strong{fgeo}'s documentation.}
\usage{
fgeo_index(.f, nm, ...)
}
\arguments{
\item{.f}{A function to map to each core-package of fgeo.}

\item{nm}{A name for the column storing the results of \code{.f}.}

\item{...}{Additional arguments passed to \code{.f}.}

\item{keep_reexported}{Logical; \code{TRUE} includes reexported functions.}
}
\description{
Create a table of some component of \strong{fgeo}'s documentation.
}
\keyword{internal}
