% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fgeo_link.R
\name{fgeo_link}
\alias{fgeo_link}
\title{Creates a functional link when used with DT table.}
\usage{
fgeo_link(pkg, fun = NULL)
}
\arguments{
\item{pkg, fun}{String; the name of a package (\code{pkg}) and function (\code{fun}).}
}
\description{
Internal function. Exported to use in vignettes.
}
\examples{
\dontrun{
fgeo::fgeo_index_packages() \%>\%
  dplyr::mutate(package = fgeo::fgeo_link(package)) \%>\%
  DT::datatable(escape = FALSE)
}
}
\keyword{internal}
