% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_objects.R
\name{fgeo_objects}
\alias{fgeo_objects}
\alias{fgeo_functions}
\alias{fgeo_data}
\title{Overview fgeo's data and functions.}
\usage{
fgeo_functions()

fgeo_data()
}
\value{
A dataframe.
}
\description{
Overview fgeo's data and functions.
}
\examples{
head(fgeo_functions())

head(fgeo_data())
}
