% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_pkgs.R
\name{fgeo_package_deps}
\alias{fgeo_package_deps}
\title{Find fgeo packages that depend on a given package.}
\usage{
fgeo_package_deps(pkg, root = "../", fgeo_pkgs = fgeo_core())
}
\arguments{
\item{pkg}{String giving the name of one package which dependencies you want
to find.}

\item{root}{Sting. Assumes all fgeo package have the same parent directory.}

\item{fgeo_pkgs}{Character string giving the names of fgeo packages.}
}
\value{
A character vector.
}
\description{
Useful to find which packages may have broken after refactoring on package
(the one you would want to pass to the argument \code{pkg}).
}
\examples{
# Won't because it depends on local directories.
\dontrun{
fgeo_package_deps("fgeo.tool")
}
}
