% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_packages.R
\name{fgeo_packages}
\alias{fgeo_packages}
\title{All and core \strong{fgeo} packages.}
\usage{
fgeo_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Logical. Should the output include fgeo?}
}
\value{
A string.
}
\description{
All and core \strong{fgeo} packages.
}
