% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_pkgs.R
\name{fgeo_pkgs}
\alias{fgeo_pkgs}
\alias{fgeo_core}
\title{All and core \strong{fgeo} packages.}
\usage{
fgeo_pkgs(include_self = TRUE)

fgeo_core()
}
\value{
A string.
}
\description{
All and core \strong{fgeo} packages.
}
\examples{
fgeo_pkgs()
fgeo_pkgs(FALSE)
fgeo_core()
}
\concept{functions for developers}
\keyword{internal}
