% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{fgeo_update}
\alias{fgeo_update}
\title{Update fgeo packages}
\usage{
fgeo_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
fgeo packages.}
}
\description{
This will check to see if all fgeo packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
fgeo_update()
}
}
