% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.R
\name{map}
\alias{map}
\title{map: Map ForestGEO's data.}
\section{Installation}{
\preformatted{# install.packages("remotes")
remotes::install_github("forestgeo/map")
}

\href{https://forestgeo.github.io/map/}{Learn more about map}
}

