% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-fgeo-analyze.R, R/reexport-fgeo-map.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{basal_area}
\alias{abundance_byyr}
\alias{basal_area_byyr}
\alias{fgeo_habitat}
\alias{fgeo_topography}
\alias{growth_ctfs}
\alias{mortality_ctfs}
\alias{recruitment_ctfs}
\alias{to_df}
\alias{tt_test}
\alias{autoplot}
\alias{autoplot_by_species}
\alias{elev}
\alias{header_dbh_bubbles}
\alias{header_tag_status}
\alias{plot_dbh_bubbles_by_quadrat}
\alias{plot_tag_status_by_subquadrat}
\alias{sp}
\alias{sp_elev}
\alias{theme_dbh_bubbles}
\alias{theme_default}
\alias{theme_tag_status}
\alias{vft_1quad}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{fgeo.analyze}{\code{\link[fgeo.analyze]{basal_area}}, \code{\link[fgeo.analyze]{abundance_byyr}}, \code{\link[fgeo.analyze]{basal_area_byyr}}, \code{\link[fgeo.analyze]{fgeo_habitat}}, \code{\link[fgeo.analyze]{fgeo_topography}}, \code{\link[fgeo.analyze]{growth_ctfs}}, \code{\link[fgeo.analyze]{mortality_ctfs}}, \code{\link[fgeo.analyze]{recruitment_ctfs}}, \code{\link[fgeo.analyze]{to_df}}, \code{\link[fgeo.analyze]{tt_test}}}

  \item{fgeo.map}{\code{\link[fgeo.map]{autoplot}}, \code{\link[fgeo.map]{autoplot_by_species}}, \code{\link[fgeo.map]{elev}}, \code{\link[fgeo.map]{header_dbh_bubbles}}, \code{\link[fgeo.map]{header_tag_status}}, \code{\link[fgeo.map]{plot_dbh_bubbles_by_quadrat}}, \code{\link[fgeo.map]{plot_tag_status_by_subquadrat}}, \code{\link[fgeo.map]{sp}}, \code{\link[fgeo.map]{sp_elev}}, \code{\link[fgeo.map]{theme_dbh_bubbles}}, \code{\link[fgeo.map]{theme_default}}, \code{\link[fgeo.map]{theme_tag_status}}, \code{\link[fgeo.map]{vft_1quad}}}
}}

