% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.R
\name{soilkrig}
\alias{soilkrig}
\title{soilkrig: Krig soil data as per John et al. (2007).}
\section{Installation}{
\preformatted{# install.packages("remotes")
remotes::install_github("forestgeo/soilkrig")
}

\href{https://forestgeo.github.io/soilkrig/}{Learn more about soilkrig}
}

