% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_packages.R
\name{table_core}
\alias{table_core}
\title{Create a table of \strong{fgeo}'s package-names and titles.}
\usage{
table_core()
}
\description{
Create a table of \strong{fgeo}'s package-names and titles.
}
\section{Acknowledgments}{

\href{https://goo.gl/Qzp5kR}{Jim Hester advised} and \href{http://bit.ly/2H8bwaj}{Nathan Werth advised} helped develop the implementation details of
this function.
}

