% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo-package.R
\docType{package}
\name{fgeo-package}
\alias{fgeo}
\alias{fgeo-package}
\title{fgeo: Analyze Forest Diversity and Dynamics}
\description{
'fgeo' allows you to install, load, and access
    the documentation of multiple packages to analyze forest diversity and
    dynamics. The collection includes packages to manipulate and plot
    ForestGEO data, and to do common analyses including abundance,
    demography, and species-habitats associations
    (<http://www.forestgeo.si.edu/>).
}
\seealso{
Useful links:
\itemize{
  \item \url{http://forestgeo.github.io/fgeo}
  \item \url{https://github.com/forestgeo/fgeo}
  \item Report bugs at \url{https://github.com/forestgeo/fgeo/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} (0000-0002-1986-7988) [contractor]

Authors:
\itemize{
  \item Gabriel Arellano \email{gabriel.arellano.torres@gmail.com} [reviewer]
  \item Richard  Condit \email{richardcondit@gmail.com}
  \item Stuart Davies \email{daviess@si.edu} [reviewer]
  \item Matteo Detto \email{teoparaglider@gmail.com}
  \item Erika Gonzalez-Akre \email{GonzalezEB@si.edu} [reviewer]
  \item Pamela Hall \email{phall@alum.mit.edu}
  \item Kyle Harms \email{ kharms@lsu.edu}
  \item Valentine Herrmann \email{HerrmannV@si.edu} (0000-0002-4519-481X) [reviewer]
  \item David Kenfack \email{KenfackD@si.edu } [reviewer]
  \item Suzanne Lao \email{laoz@si.edu} [reviewer]
  \item Sean McMahon \email{mcmahons@si.edu} [reviewer]
  \item Sabrina Russo \email{srusso2@unl.edu} [reviewer]
  \item Kristina Teixeira \email{TeixeiraK@si.edu} [reviewer]
  \item Graham Zemunik \email{grah.zem@gmail.com} [reviewer]
  \item Daniel Zuleta \email{dfzuleta@gmail.com} [reviewer]
}

Other contributors:
\itemize{
  \item Aaron Hogan \email{jamesaaronhogan@gmail.com} (0000-0001-9806-3074) [reviewer]
  \item Bier Kraichak \email{ekraichak@gmail.com} [reviewer]
  \item Lauren Krizel \email{KrizelL@si.edu} [reviewer]
  \item Haley Overstreet \email{OverstreetH@si.edu} [reviewer]
  \item Cara Scalpone \email{ScalponeC@si.edu} (0000-0001-8448-2147) [reviewer]
  \item CTFS-ForestGEO \email{ForestGEO@si.edu} [copyright holder, funder]
}

}
\keyword{internal}
