% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.R
\docType{package}
\name{fgeo-package}
\alias{fgeo}
\alias{fgeo-package}
\title{fgeo: Easily Install and Load Packages of ForestGEO}
\description{
'fgeo' installs and load multiple R packages of ForestGEO in
a single step. Each individual package is a self contained module. Together,
all the packages in 'fgeo' provide a wide renge of tools to analyze forest
diversity and dynamics. The modular structure of 'fgeo' makes it easy to add
and maintain more tools as they become available. Learn more about the
'fgeo' package at <http://forestgeo.github.io/fgeo>; and learn
more about ForestGEO at <http://www.forestgeo.si.edu/>.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://forestgeo.github.io/fgeo}
  \item \url{https://github.com/forestgeo/fgeo}
  \item Report bugs at \url{https://github.com/forestgeo/fgeo/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} [contractor]

Other contributors:
\itemize{
  \item CTFS-ForestGEO \email{ForestGEO@si.edu} [copyright holder, funder]
}

}
\keyword{internal}
