% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.R
\docType{package}
\name{fgeo-package}
\alias{fgeo}
\alias{fgeo-package}
\title{fgeo: Easily Install and Load Multiple Packages}
\description{
'fgeo' installs and loads multiple R packages of ForestGEO in
    a single step. Each individual package is a self-contained module. Together,
    all the packages in 'fgeo' provide a wide range of tools to analyze forest
    diversity and dynamics. The modular structure of 'fgeo' makes it easy to add 
    and maintain more tools as they become available. Learn more about the 
    'fgeo' package at <http://forestgeo.github.io/fgeo>; and learn
    more about ForestGEO at <http://www.forestgeo.si.edu/>.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://forestgeo.github.io/fgeo}
  \item \url{https://github.com/forestgeo/fgeo}
  \item Report bugs at \url{https://github.com/forestgeo/fgeo/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} [contractor]

Other contributors:
\itemize{
  \item Gabriel Arellano \email{gabriel.arellano.torres@gmail.com} [reviewer]
  \item Stuart Davies \email{daviess@si.edu} [reviewer]
  \item Lauren Krizel \email{KrizelL@si.edu} [reviewer]
  \item Suzanne Lao \email{laoz@si.edu} [reviewer]
  \item Sean McMahon \email{mcmahons@si.edu} [reviewer]
  \item Haley Overstreet \email{OverstreetH@si.edu} [reviewer]
  \item Kristina Teixeira \email{TeixeiraK@si.edu} [reviewer]
  \item CTFS-ForestGEO \email{ForestGEO@si.edu} [copyright holder, funder]
}

}
\keyword{internal}
