% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.R
\name{fgeo.demography}
\alias{fgeo.demography}
\title{fgeo.demography: Analyze recruitment, mortality and growth.}
\section{Installation}{

To install \strong{fgeo.demography} run the following code in R:\preformatted{missing_remotes <- !require(remotes)
if (missing_remotes) {
  install.packages("remotes")
}
remotes::install_github("forestgeo/fgeo.demography")
}
\itemize{
\item To learn the details about how to install packages from GitHub, read
\href{https://goo.gl/dQKEeg}{this blog post}.
\item \href{https://forestgeo.github.io/fgeo.demography/}{Go to the website of fgeo.demography}
}
}

