% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.R
\name{fgeo.habitat}
\alias{fgeo.habitat}
\title{fgeo.habitat: Habitat analyses.}
\section{Installation}{
\preformatted{# install.packages("remotes")
remotes::install_github("forestgeo/fgeo.habitat")
}

\href{https://forestgeo.github.io/fgeo.habitat/}{Learn more about fgeo.habitat}
}

