% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.R
\name{fgeo.map}
\alias{fgeo.map}
\title{fgeo.map: Map ForestGEO's data.}
\section{Installation}{

To install \strong{fgeo.map} run the following code in R:\preformatted{missing_remotes <- !require(remotes)
if (missing_remotes) {
  install.packages("remotes")
}
remotes::install_github("forestgeo/fgeo.map")
}
\itemize{
\item To learn the details about how to install packages from GitHub, read
\href{https://goo.gl/dQKEeg}{this blog post}.
\item \href{https://forestgeo.github.io/fgeo.map/}{Go to the website of fgeo.map}
}
}

