% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.R
\name{fgeo.tool}
\alias{fgeo.tool}
\title{fgeo.tool: Utility functions of ForestGEO}
\section{Installation}{
\preformatted{# install.packages("remotes")
remotes::install_github("forestgeo/fgeo.tool")
}

\href{https://forestgeo.github.io/fgeo.tool/}{Learn more about fgeo.tool}
}

