% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_dependencies.R
\name{fgeo_dependencies}
\alias{fgeo_dependencies}
\title{List package dependencies (imports and suggests) of \strong{fgeo}.}
\usage{
fgeo_dependencies(matches = NULL, include_self = TRUE,
  section = c("Imports", "Suggests"))
}
\arguments{
\item{matches}{String to \code{grep()} the output.}

\item{include_self}{Include fgeo in the list?}

\item{section}{Either "Imports" or "Suggests".}
}
\value{
A string.
}
\description{
List package dependencies (imports and suggests) of \strong{fgeo}.
}
\concept{functions for developers}
