% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_help.R
\name{fgeo_help}
\alias{fgeo_help}
\alias{fgeo_help_addin}
\title{Search \strong{fgeo} for documentation matching a given character string.}
\usage{
fgeo_help(pattern = "", package = fgeo::fgeo_core())

fgeo_help_addin()
}
\arguments{
\item{pattern}{A character string to be matched in the name, alias or title
of a topic's documentation. Defaults to matching anything, which retrieves
all the documentation of \strong{fgeo} packages.}

\item{package}{A character string giving the name or one or more
core-packages of \strong{fgeo}.}
}
\value{
The results are returned in a list object of class "hsearch", which
has a print method for nicely formatting the results of the query.
}
\description{
\itemize{
\item \code{fgeo_help()} finds all \strong{fgeo} help files. You can refine
your search directly on the viewer panel of RStudio or by passing a string
to \code{fgeo_help()} (e.g. \code{fgeo("datasets")}).
\item \code{fgeo_help_adin()} runs \code{fgeo_help()} from the RStudio Addins menu or by
associating a keyboard shortcut to it (see Tools > Modify Keyboard
Shortcuts...), for example, Ctr/Cmd + Shift + Alt + F1.
}
}
\examples{
if (interactive()) fgeo_help()
if (interactive()) fgeo_help("fgeo", package = "fgeo")

results <- fgeo_help("fgeo", package = "fgeo.tool")
if (interactive()) results
str(results)
}
\seealso{
\code{\link[utils:help.search]{utils::help.search()}}, \code{\link[fgeo:fgeo_core]{fgeo::fgeo_core()}}.
}
