% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_index.R
\name{fgeo_index}
\alias{fgeo_index}
\alias{fgeo_index_functions}
\alias{fgeo_index_packages}
\alias{fgeo_index_datasets}
\title{Index fgeo's packages, datasets and functions.}
\usage{
fgeo_index(.f, nm, ...)

fgeo_index_functions(keep_reexported = FALSE)

fgeo_index_packages()

fgeo_index_datasets()
}
\arguments{
\item{.f}{A function to map to each core-package of fgeo.}

\item{nm}{A name for the column storing the results of \code{.f}.}

\item{...}{Additional arguments passed to \code{.f}.}

\item{keep_reexported}{Logical; \code{TRUE} includes reexported functions.}
}
\value{
A dataframe.
}
\description{
\code{fgeo_index_packages()} creates a table of all packages of the \strong{fgeo}
meta-package. It is searchable by package name and by whatever keyword
appears in their titles, and links to th e website of each package.
\code{fgeo_index_datasets()} and \code{fgeo_index_functions()} create tables of all
datasets and functions of the \strong{fgeo} meta-package. It is searchable by
dataset and function name and includes links to the reference page of the
respective package-website (a link to the specific documentation of each
dataset and function would be better but this feature is not yet
available).
}
\section{Acknowledgements}{

\href{https://goo.gl/Qzp5kR}{Jim Hester adviced} on implementation details of
\code{fgeo_index_functions()}; and \href{http://bit.ly/2H8bwaj}{Nathan Werth adviced}
on implementation details of \code{fgeo_index_datasets()}.
}

\examples{
head(fgeo_index_functions())
head(fgeo_index_packages())
}
