% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_index.R
\name{fgeo_index}
\alias{fgeo_index}
\alias{fgeo_index_functions}
\alias{fgeo_index_packages}
\title{Index fgeo's elements.}
\usage{
fgeo_index(.f, nm, ...)

fgeo_index_functions()

fgeo_index_packages()
}
\arguments{
\item{.f}{A function to map to each core package of fgeo.}

\item{nm}{A name for the column storing the results of \code{.f}.}

\item{...}{Additional arguments passed to \code{.f}.}
}
\value{
A dataframe.
}
\description{
\code{fgeo_index()} created a table of some component of fgeo for each core
package of fgeo. The other functions are convenient wrappers.
}
\section{Acknowledgements}{

Thanks to Jim Hester for \href{https://goo.gl/Qzp5kR}{advice} on an implementation
detail.
}

\examples{
head(fgeo_index_functions())
head(fgeo_index_packages())
}
