% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_packages.R
\name{fgeo_packages}
\alias{fgeo_packages}
\title{Create a table of \strong{fgeo}'s package-names and titles.}
\usage{
fgeo_packages()
}
\description{
Create a table of \strong{fgeo}'s package-names and titles.
}
\section{Acknowledgments}{

\href{https://goo.gl/Qzp5kR}{Jim Hester advised} on implementation details of
\code{fgeo_index_functions()}; and \href{http://bit.ly/2H8bwaj}{Nathan Werth advised}
on implementation details of \code{fgeo_index_datasets()}.
}

\keyword{internal}
