% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fgeo_packages}
\alias{fgeo_packages}
\title{List all packages in fgeo}
\usage{
fgeo_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include fgeo in the list?}
}
\description{
List all packages in fgeo
}
\examples{
fgeo_packages()
}
