% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_pkgs.R
\name{fgeo_packages}
\alias{fgeo_packages}
\alias{fgeo_imports}
\alias{fgeo_suggests}
\title{List packages that fgeo imports and suggests.}
\usage{
fgeo_packages(include_self = TRUE, section = c("Imports", "Suggests"))

fgeo_imports(include_self = TRUE)

fgeo_suggests()
}
\arguments{
\item{include_self}{Include fgeo in the list?}

\item{section}{Either "Imports" or "Suggests".}
}
\description{
List packages that fgeo imports and suggests.
}
\examples{
fgeo_packages(TRUE, "Imports")
# Same
fgeo_imports()

fgeo_suggests()
}
