% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_packages.R
\name{fgeo_packages}
\alias{fgeo_packages}
\alias{fgeo_core}
\title{All and core \strong{fgeo} packages.}
\usage{
fgeo_packages(include_self = TRUE)

fgeo_core()
}
\value{
A string.
}
\description{
All and core \strong{fgeo} packages.
}
\examples{
fgeo_packages()
fgeo_packages(FALSE)
fgeo_core()
}
\concept{functions for developers}
\keyword{internal}
