% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_packages.R
\name{fgeo_packages}
\alias{fgeo_packages}
\title{Names of the core fgeo packages.}
\usage{
fgeo_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Logical. Should the output include fgeo?}
}
\value{
A string.
}
\description{
Names of the core fgeo packages.
}
\examples{
fgeo_packages()
fgeo_packages(include_self = FALSE)
}
\seealso{
Other functions to explore fgeo: \code{\link{fgeo_browse}},
  \code{\link{fgeo_help}}
}
\concept{functions to explore fgeo}
