% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo_update.R
\name{fgeo_update}
\alias{fgeo_update}
\title{Update \strong{fgeo}.}
\usage{
fgeo_update()
}
\description{
Shortcut for \code{remotes::install_github("forestgeo/fgeo")}.
}
\examples{
\dontrun{
fgeo_update()
}
}
