% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-fgeo-analyze.R, R/reexport-fgeo-map.R,
%   R/reexport-tool.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{basal_area}
\alias{abundance_byyr}
\alias{basal_area_byyr}
\alias{fgeo_habitat}
\alias{fgeo_topography}
\alias{growth_ctfs}
\alias{mortality_ctfs}
\alias{recruitment_ctfs}
\alias{to_df}
\alias{tt_test}
\alias{autoplot}
\alias{autoplot_by_species}
\alias{elev}
\alias{header_dbh_bubbles}
\alias{header_tag_status}
\alias{plot_dbh_bubbles_by_quadrat}
\alias{plot_tag_status_by_subquadrat}
\alias{sp}
\alias{sp_elev}
\alias{theme_dbh_bubbles}
\alias{theme_default}
\alias{theme_tag_status}
\alias{vft_1quad}
\alias{add_col_row}
\alias{add_col_row2}
\alias{add_count}
\alias{add_hectindex}
\alias{add_id}
\alias{add_index}
\alias{add_lxly}
\alias{add_quad}
\alias{add_qxqy}
\alias{add_status_tree}
\alias{add_subquad}
\alias{arrange}
\alias{as_censuses}
\alias{as_tibble}
\alias{check_crucial_names}
\alias{contains}
\alias{count}
\alias{csv_list}
\alias{delim_list}
\alias{detect_if_group}
\alias{detect_insensitive}
\alias{drop_if_na}
\alias{drop_status}
\alias{drop_twice_dead}
\alias{ends_with}
\alias{enexpr}
\alias{enquo}
\alias{enquos}
\alias{ensym}
\alias{everything}
\alias{expr}
\alias{exprs}
\alias{extract_gridsize}
\alias{extract_insensitive}
\alias{extract_plotdim}
\alias{fgeo_elevation}
\alias{fill_na}
\alias{filter}
\alias{flag_if}
\alias{flag_if_group}
\alias{group_by}
\alias{guess_plotdim}
\alias{gxgy_to_hectindex}
\alias{gxgy_to_index}
\alias{gxgy_to_lxly}
\alias{gxgy_to_qxqy}
\alias{gxgy_to_rowcol}
\alias{hide_data_of_class}
\alias{index_to_gxgy}
\alias{index_to_rowcol}
\alias{is_duplicated}
\alias{is_multiple}
\alias{last_col}
\alias{list_csv}
\alias{list_df}
\alias{lookup}
\alias{matches}
\alias{mutate}
\alias{name_dfs}
\alias{nms_try_rename}
\alias{num_range}
\alias{one_of}
\alias{pick}
\alias{pick_dbh_max}
\alias{pick_dbh_min}
\alias{pick_dbh_over}
\alias{pick_dbh_under}
\alias{pick_main_stem}
\alias{pick_main_stemid}
\alias{pick_plotname}
\alias{pick_recensus}
\alias{pick_status}
\alias{pick_top}
\alias{quo}
\alias{quo_name}
\alias{quos}
\alias{rda_list}
\alias{rdata_list}
\alias{rds_list}
\alias{read_taxa}
\alias{read_vft}
\alias{read_with}
\alias{recode_subquad}
\alias{rename_matches}
\alias{rowcol_to_index}
\alias{sanitize_taxa}
\alias{sanitize_vft}
\alias{select}
\alias{show_data_of_class}
\alias{starts_with}
\alias{suffix_match}
\alias{suffix_tags_beyond_edge}
\alias{summarise}
\alias{summarize}
\alias{sym}
\alias{syms}
\alias{tibble}
\alias{tool_example}
\alias{tribble}
\alias{tsv_list}
\alias{type_ensure}
\alias{type_taxa}
\alias{type_vft}
\alias{ungroup}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{fgeo.analyze}{\code{\link[fgeo.analyze]{basal_area}}, \code{\link[fgeo.analyze]{abundance_byyr}}, \code{\link[fgeo.analyze]{basal_area_byyr}}, \code{\link[fgeo.analyze]{fgeo_habitat}}, \code{\link[fgeo.analyze]{fgeo_topography}}, \code{\link[fgeo.analyze]{growth_ctfs}}, \code{\link[fgeo.analyze]{mortality_ctfs}}, \code{\link[fgeo.analyze]{recruitment_ctfs}}, \code{\link[fgeo.analyze]{to_df}}, \code{\link[fgeo.analyze]{tt_test}}}

  \item{fgeo.map}{\code{\link[fgeo.map]{autoplot}}, \code{\link[fgeo.map]{autoplot_by_species}}, \code{\link[fgeo.map]{elev}}, \code{\link[fgeo.map]{header_dbh_bubbles}}, \code{\link[fgeo.map]{header_tag_status}}, \code{\link[fgeo.map]{plot_dbh_bubbles_by_quadrat}}, \code{\link[fgeo.map]{plot_tag_status_by_subquadrat}}, \code{\link[fgeo.map]{sp}}, \code{\link[fgeo.map]{sp_elev}}, \code{\link[fgeo.map]{theme_dbh_bubbles}}, \code{\link[fgeo.map]{theme_default}}, \code{\link[fgeo.map]{theme_tag_status}}, \code{\link[fgeo.map]{vft_1quad}}}

  \item{fgeo.tool}{\code{\link[fgeo.tool]{add_col_row}}, \code{\link[fgeo.tool]{add_col_row2}}, \code{\link[fgeo.tool]{add_count}}, \code{\link[fgeo.tool]{add_hectindex}}, \code{\link[fgeo.tool]{add_id}}, \code{\link[fgeo.tool]{add_index}}, \code{\link[fgeo.tool]{add_lxly}}, \code{\link[fgeo.tool]{add_quad}}, \code{\link[fgeo.tool]{add_qxqy}}, \code{\link[fgeo.tool]{add_status_tree}}, \code{\link[fgeo.tool]{add_subquad}}, \code{\link[fgeo.tool]{arrange}}, \code{\link[fgeo.tool]{as_censuses}}, \code{\link[fgeo.tool]{as_tibble}}, \code{\link[fgeo.tool]{check_crucial_names}}, \code{\link[fgeo.tool]{contains}}, \code{\link[fgeo.tool]{count}}, \code{\link[fgeo.tool]{csv_list}}, \code{\link[fgeo.tool]{delim_list}}, \code{\link[fgeo.tool]{detect_if_group}}, \code{\link[fgeo.tool]{detect_insensitive}}, \code{\link[fgeo.tool]{drop_if_na}}, \code{\link[fgeo.tool]{drop_status}}, \code{\link[fgeo.tool]{drop_twice_dead}}, \code{\link[fgeo.tool]{ends_with}}, \code{\link[fgeo.tool]{enexpr}}, \code{\link[fgeo.tool]{enquo}}, \code{\link[fgeo.tool]{enquos}}, \code{\link[fgeo.tool]{ensym}}, \code{\link[fgeo.tool]{everything}}, \code{\link[fgeo.tool]{expr}}, \code{\link[fgeo.tool]{exprs}}, \code{\link[fgeo.tool]{extract_gridsize}}, \code{\link[fgeo.tool]{extract_insensitive}}, \code{\link[fgeo.tool]{extract_plotdim}}, \code{\link[fgeo.tool]{fgeo_elevation}}, \code{\link[fgeo.tool]{fill_na}}, \code{\link[fgeo.tool]{filter}}, \code{\link[fgeo.tool]{flag_if}}, \code{\link[fgeo.tool]{flag_if_group}}, \code{\link[fgeo.tool]{group_by}}, \code{\link[fgeo.tool]{guess_plotdim}}, \code{\link[fgeo.tool]{gxgy_to_hectindex}}, \code{\link[fgeo.tool]{gxgy_to_index}}, \code{\link[fgeo.tool]{gxgy_to_lxly}}, \code{\link[fgeo.tool]{gxgy_to_qxqy}}, \code{\link[fgeo.tool]{gxgy_to_rowcol}}, \code{\link[fgeo.tool]{hide_data_of_class}}, \code{\link[fgeo.tool]{index_to_gxgy}}, \code{\link[fgeo.tool]{index_to_rowcol}}, \code{\link[fgeo.tool]{is_duplicated}}, \code{\link[fgeo.tool]{is_multiple}}, \code{\link[fgeo.tool]{last_col}}, \code{\link[fgeo.tool]{list_csv}}, \code{\link[fgeo.tool]{list_df}}, \code{\link[fgeo.tool]{lookup}}, \code{\link[fgeo.tool]{matches}}, \code{\link[fgeo.tool]{mutate}}, \code{\link[fgeo.tool]{name_dfs}}, \code{\link[fgeo.tool]{nms_try_rename}}, \code{\link[fgeo.tool]{num_range}}, \code{\link[fgeo.tool]{one_of}}, \code{\link[fgeo.tool]{pick}}, \code{\link[fgeo.tool]{pick_dbh_max}}, \code{\link[fgeo.tool]{pick_dbh_min}}, \code{\link[fgeo.tool]{pick_dbh_over}}, \code{\link[fgeo.tool]{pick_dbh_under}}, \code{\link[fgeo.tool]{pick_main_stem}}, \code{\link[fgeo.tool]{pick_main_stemid}}, \code{\link[fgeo.tool]{pick_plotname}}, \code{\link[fgeo.tool]{pick_recensus}}, \code{\link[fgeo.tool]{pick_status}}, \code{\link[fgeo.tool]{pick_top}}, \code{\link[fgeo.tool]{quo}}, \code{\link[fgeo.tool]{quo_name}}, \code{\link[fgeo.tool]{quos}}, \code{\link[fgeo.tool]{rda_list}}, \code{\link[fgeo.tool]{rdata_list}}, \code{\link[fgeo.tool]{rds_list}}, \code{\link[fgeo.tool]{read_taxa}}, \code{\link[fgeo.tool]{read_vft}}, \code{\link[fgeo.tool]{read_with}}, \code{\link[fgeo.tool]{recode_subquad}}, \code{\link[fgeo.tool]{rename_matches}}, \code{\link[fgeo.tool]{rowcol_to_index}}, \code{\link[fgeo.tool]{sanitize_taxa}}, \code{\link[fgeo.tool]{sanitize_vft}}, \code{\link[fgeo.tool]{select}}, \code{\link[fgeo.tool]{show_data_of_class}}, \code{\link[fgeo.tool]{starts_with}}, \code{\link[fgeo.tool]{suffix_match}}, \code{\link[fgeo.tool]{suffix_tags_beyond_edge}}, \code{\link[fgeo.tool]{summarise}}, \code{\link[fgeo.tool]{summarize}}, \code{\link[fgeo.tool]{sym}}, \code{\link[fgeo.tool]{syms}}, \code{\link[fgeo.tool]{tibble}}, \code{\link[fgeo.tool]{tool_example}}, \code{\link[fgeo.tool]{tribble}}, \code{\link[fgeo.tool]{tsv_list}}, \code{\link[fgeo.tool]{type_ensure}}, \code{\link[fgeo.tool]{type_taxa}}, \code{\link[fgeo.tool]{type_vft}}, \code{\link[fgeo.tool]{ungroup}}}
}}

